move Location from shared to components
