+++ src/App.jsx
@@ -26,7 +26,7 @@ const screens = {
 screens.init = require('./components/Init').default;
 screens.home = require('./components/Home').default;
 screens.about = require('./components/About').default;
-screens.location = require('./shared/Location').default;
+screens.location = require('./components/Location').default;
 screens.soil = require('./components/Soil').default;
 screens.covercrop = require('./components/CoverCrop').CoverCropFirst;
 screens.covercrop2 = require('./components/CoverCrop').CoverCropSecond;
